move package code into distro and inventory/serial
